fix: 인터페이스 정리 #122
